Refactor: Remove unnecessary enum cast in method call
Simplified the call to CambiarEstatusAsync by removing the cast to (int) for Enums.EstatusMuestreo.MóduloInicialReglas. The method now directly accepts the enum type, improving code readability and reducing potential casting errors.

+++ Application/Features/Operacion/Resultados/Comands/Acumulacion/EnviarInicialReglasCommand.cs
@@ -23,7 +23,7 @@ namespace Application.Features.Operacion.Resultados.Comands.Acumulacion
         {
             if (request.Muestreos.Any())
             {
-                bool actualizados = await _muestreoRepository.CambiarEstatusAsync((int)Enums.EstatusMuestreo.MóduloInicialReglas, request.Muestreos);
+                bool actualizados = await _muestreoRepository.CambiarEstatusAsync(Enums.EstatusMuestreo.MóduloInicialReglas, request.Muestreos);
                 return new Response<bool> { Succeded = actualizados };
             }
             else
